+++ src/main/java/com/wareflow/buildify/domain/auth/login/security/CustomUserDetails.java
@@ -1,8 +1,6 @@
 package com.wareflow.buildify.domain.auth.login.security;
 
-import lombok.Builder;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
+import lombok.*;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -11,25 +9,17 @@ import java.util.Collection;
 import java.util.List;
 
 @Getter
-@RequiredArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED) // ✅ 이거만 남김
+@AllArgsConstructor
+@Builder
 public class CustomUserDetails implements UserDetails {
 
-    private final String id;
-    private final String password;
-    private final String role;
+    private String id;
+    private String password;
+    private String role;
 
     private String clientId;    // ✅ 추가 (User일 경우)
     private String adminNumber; // ✅ 추가 (Admin일 경우)
-
-    @Builder
-    public CustomUserDetails(String id, String password, String role,
-                             String clientId, String adminNumber) {
-        this.id = id;
-        this.password = password;
-        this.role = role;
-        this.clientId = clientId;
-        this.adminNumber = adminNumber;
-    }
 
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
